backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
requests==2.32.3
pillow==12.1.0
transliterate==1.10.2
